Refactored database into more files
