+++ server/database/constants.js
@@ -1,0 +1,9 @@
+module.exports = {
+    ITEM_TABLE: 'item_table',
+    LIST_TABLE: 'list_table',
+    CATEGORY_TABLE: 'category_table',
+    META_TABLE: 'meta_table',
+    CATEGORY_NONE: 'Uncategorized',
+
+    CURRENT_VERSION: 2,
+};
